final commit

+++ app/Http/Controllers/ProjectController.php
@@ -2,27 +2,31 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 use App\Models\Project;
 use App\Models\User;
-use Illuminate\Support\Facades\Gate;
+use App\Traits\Access;
+use Illuminate\Http\Request;
 
 class ProjectController extends Controller
 {
+    use Access;
+
     /**
      * Display a listing of the resource.
      */
+    public function __construct()
+    {
+        $this->authorizeResource(Project::class);
+    }
+
     public function index()
     {
         //check user authorized
-        $this->isUserHavePermission('view-project');
-        $response = Gate::inspect('check-project-access');
-        if (!$response->allowed()) {
+        if (! $this->isUserHaveAccessToProject(null)) {
             $projects = User::findOrFail(Auth()->user()->id)->projects;
         } else {
             $projects = Project::all();
         }
-
 
         return response()->json($projects);
     }
@@ -40,27 +44,24 @@ class ProjectController extends Controller
      */
     public function store(Request $request)
     {
-        //check user autherized
-        $this->isUserHavePermission('create-project');
-
-
         $validated = $request->validate(['project_name' => 'required|max:250|min:1']);
         $project = new Project();
-        $project->project_name = $validated["project_name"];
+        $project->project_name = $validated['project_name'];
         $project->save();
+
         return response()->json($project);
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(string $id)
+    public function show(Project $project)
     {
         //check authorized or not
-        $this->isUserHavePermission('view-project');
-        $this->isUserHaveAccessToProject($id);
+        if (! $this->isUserHaveAccessToProject($project->id)) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
 
-        $project = Project::findOrFail($id);
         return response()->json($project);
     }
 
@@ -75,14 +76,15 @@ class ProjectController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, string $id)
+    public function update(Request $request, Project $project)
     {
         //check authorized or not
-        $this->isUserHavePermission('update-project');
-        $this->isUserHaveAccessToProject($id);
+        if (! $this->isUserHaveAccessToProject($project->id)) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
 
         $validated = $request->validate(['project_name' => 'required|max:250|min:1']);
-        $project = project::findOrFail($id);
+        $project = project::findOrFail($project->id);
         $project->project_name = $validated['project_name'];
         $project->save();
 
@@ -92,31 +94,20 @@ class ProjectController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(Project $project)
     {
         //check authorized or not
-        $this->isUserHavePermission('delete-project');
-        $this->isUserHaveAccessToProject($id);
-
-        $project = Project::findOrFail($id);
+        if (! $this->isUserHaveAccessToProject($project->id)) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
+        $project = Project::findOrFail($project->id);
         $project->delete();
-        // dd($id);
+
         return response()->json($project);
     }
 
-    public function isUserHavePermission($permission)
+    public function hello()
     {
-        $response = Gate::inspect($permission);
-        if (!$response->allowed()) {
-            return response()->json(['error' => 'Unauthorized'], 403);
-        }
-    }
-
-    public function isUserHaveAccessToProject($projectId)
-    {
-        $response = Gate::inspect('check-project-access', [$projectId]);
-        if (!$response->allowed()) {
-            return response()->json(['error' => 'Unauthorized'], 403);
-        }
+        return 'hello';
     }
 }
